Split large test file
